refactor(find): moved out of extended directory
BREAKING CHANGE: `find` has been moved out of `extended` folder and under `operator` folder directly
BREAKING CHANGE: `findIndex` has been moved out of `extended` folder and under `operator` folder directly
