+++ src/Rx.KitchenSink.ts
@@ -70,8 +70,8 @@ import './add/operator/exhaust';
 import './add/operator/exhaustMap';
 import './add/operator/expand';
 import './add/operator/filter';
-import './add/operator/extended/find';
-import './add/operator/extended/findIndex';
+import './add/operator/find';
+import './add/operator/findIndex';
 import './add/operator/finally';
 import './add/operator/first';
 import './add/operator/groupBy';
