--- src/add/operator/extended/find.ts
@@ -1,7 +1,0 @@
-import {Observable} from '../../../Observable';
-import {find} from '../../../operator/extended/find';
-import {KitchenSinkOperators} from '../../../Rx.KitchenSink';
-const observableProto = (<KitchenSinkOperators<any>>Observable.prototype);
-observableProto.find = find;
-
-export var _void: void;
