--- src/add/operator/extended/findIndex.ts
@@ -1,7 +1,0 @@
-import {Observable} from '../../../Observable';
-import {findIndex} from '../../../operator/extended/findIndex';
-import {KitchenSinkOperators} from '../../../Rx.KitchenSink';
-const observableProto = (<KitchenSinkOperators<any>>Observable.prototype);
-observableProto.findIndex = findIndex;
-
-export var _void: void;
