+++ src/operator/find.ts
@@ -1,5 +1,5 @@
 import {FindValueOperator} from './find-support';
-import {Observable} from '../../Observable';
+import {Observable} from '../Observable';
 
 export function find<T>(predicate: (value: T, index: number, source: Observable<T>) => boolean, thisArg?: any): Observable<T> {
   if (typeof predicate !== 'function') {
